custom-auth/requirements.txt
python-jose==3.2.0
boto3==1.28.23
botocore==1.31.23